Removing functions that have been put in fhi

+++ R/initialise.R
@@ -1,3 +1,10 @@
+#' Allows for InitialiseProject to manipulate files
+#' on your system
+#' @export AllowFileManipulationFromInitialiseProject
+AllowFileManipulationFromInitialiseProject <- function(){
+  PROJ$ALLOW_FILE_MANIPULATION_FROM_INITIALISE_PROJECT <- TRUE
+}
+
 #' Initialises project
 #' @param PROJHOME a
 #' @param PROJRAW a
@@ -12,87 +19,12 @@
 #' @param FINAL a
 #' @param SHARED a
 #' @export InitialiseProject
-InitialiseProject <- function(PROJHOME=NULL,
-                              PROJRAW=NULL,
-                              PROJCLEAN=NULL,
-                              PROJBAKED=NULL,
-                              PROJFINAL=NULL,
-                              PROJSHARED=NULL,
-                              HOME=NULL,
+InitialiseProject <- function(HOME=NULL,
                               RAW=NULL,
                               CLEAN=NULL,
                               BAKED=NULL,
                               FINAL=NULL,
                               SHARED=NULL){
-  if(Sys.getenv("PROJHOME")!=""){
-    HOME = PROJHOME = Sys.getenv("PROJHOME")
-  }
-  if(Sys.getenv("PROJRAW")!=""){
-    RAW = PROJRAW = Sys.getenv("PROJRAW")
-  }
-  if(Sys.getenv("PROJCLEAN")!=""){
-    CLEAN = PROJCLEAN = Sys.getenv("PROJCLEAN")
-  }
-  if(Sys.getenv("PROJBAKED")!=""){
-    BAKED = PROJBAKED = Sys.getenv("PROJBAKED")
-  }
-  if(Sys.getenv("PROJFINAL")!=""){
-    FINAL = PROJFINAL = Sys.getenv("PROJFINAL")
-  }
-  if(Sys.getenv("PROJSHARED")!=""){
-    SHARED = PROJSHARED = Sys.getenv("PROJSHARED")
-  }
-
-  if(is.null(PROJHOME) & !is.null(HOME)){
-    PROJHOME <- HOME
-  }
-  if(is.null(PROJRAW) & !is.null(RAW)){
-    PROJRAW <- RAW
-  }
-  if(is.null(PROJCLEAN) & !is.null(CLEAN)){
-    PROJCLEAN <- CLEAN
-  }
-  if(is.null(PROJBAKED) & !is.null(BAKED)){
-    PROJBAKED <- BAKED
-  }
-  if(is.null(PROJFINAL) & !is.null(FINAL)){
-    PROJFINAL <- FINAL
-  }
-  if(is.null(PROJSHARED) & !is.null(SHARED)){
-    PROJSHARED <- SHARED
-  }
-
-  if(!is.null(PROJHOME) & is.null(HOME)){
-    PROJHOME -> HOME
-  }
-  if(!is.null(PROJRAW) & is.null(RAW)){
-    PROJRAW -> RAW
-  }
-  if(!is.null(PROJCLEAN) & is.null(CLEAN)){
-    PROJCLEAN -> CLEAN
-  }
-  if(!is.null(PROJBAKED) & is.null(BAKED)){
-    PROJBAKED -> BAKED
-  }
-  if(!is.null(PROJFINAL) & is.null(FINAL)){
-    PROJFINAL -> FINAL
-  }
-  if(!is.null(PROJSHARED) & is.null(SHARED)){
-    PROJSHARED -> SHARED
-  }
-
-  RPROJ$PROJHOME = PROJHOME
-  RPROJ$PROJRAW = PROJRAW
-  RPROJ$PROJCLEAN = PROJCLEAN
-  RPROJ$PROJBAKED = PROJBAKED
-  RPROJ$PROJFINAL = PROJFINAL
-  RPROJ$PROJSHARED = PROJSHARED
-  RPROJ$HOME = HOME
-  RPROJ$RAW = RAW
-  RPROJ$CLEAN = CLEAN
-  RPROJ$BAKED = BAKED
-  RPROJ$FINAL = FINAL
-  RPROJ$SHARED = SHARED
 
   PROJ$HOME <- HOME
   PROJ$RAW <- RAW
@@ -100,17 +32,21 @@ InitialiseProject <- function(PROJHOME=NULL,
   PROJ$BAKED <- BAKED
   PROJ$FINAL <- FINAL
   PROJ$SHARED <- SHARED
-  PROJ$SHARED_TODAY <- file.path(SHARED,lubridate::today())
+  if(is.null(PROJ$SHARED)){
+    PROJ$SHARED_TODAY <- NULL
+  } else {
+    PROJ$SHARED_TODAY <- file.path(SHARED,lubridate::today())
 
-  # Delete empty folders in shared folder
-  for(f in list.files(PROJ$SHARED)){
-    f2 <- file.path(PROJ$SHARED,f)
-    if(length(list.files(f2))==0){
-      unlink(f2, recursive = T)
+    # Delete empty folders in shared folder
+    if(PROJ$ALLOW_FILE_MANIPULATION_FROM_INITIALISE_PROJECT) for(f in list.files(PROJ$SHARED)){
+      f2 <- file.path(PROJ$SHARED,f)
+      if(length(list.files(f2))==0){
+        unlink(f2, recursive = T)
+      }
     }
   }
 
-  for(i in names(PROJ)){
+  if(PROJ$ALLOW_FILE_MANIPULATION_FROM_INITIALISE_PROJECT) for(i in names(PROJ)){
     if(!is.null(PROJ[[i]])) if(!dir.exists(PROJ[[i]])) dir.create(PROJ[[i]], recursive=TRUE)
   }
